galeano tree basis point
In a fork of libwebsockets by David Galeano, the forked repo
currently begins with a single patch squashing 200+ commits from
the original libwebsockets project, mainly by Andy Green, into
a single commit "by" David Galeano, ie, removing the project
history and correct commit authorship.

To complicate matters that single commit is already modified from
the point the projects diverged, around October 2011.  This
commit rectifies that deletion of history by selecting a nearby
point in the canonical tree's history, at Oct 4th 2011, and then
applying the delta, under David's name, to align it with the
state of the original tree.

The rest of David's tree is applied on top.

Signed-off-by: Andy Green <andy.green@linaro.org>